chore: Refactor code, and check is precompile for create colision
